+++ src/components/Divider.tsx
@@ -3,7 +3,7 @@ import { Box, Divider as MuiDivider } from '@mui/material'
 const Divider = (): JSX.Element => {
   return (
     <Box pb={0}>
-      <MuiDivider />
+      <MuiDivider sx={{ my: 2 }} />
     </Box>
   )
 }
