+++ src/pages/CommandIndex/index.ts
@@ -1,4 +1,3 @@
 export * from './CommandIndex'
 export * from './commandIndexHelpers'
-export * from './CommandIndexTable'
 export * from './ExecuteButton'
