+++ src/pages/RequestView/RequestView.tsx
@@ -7,12 +7,12 @@ import {
   Box,
   Button,
   CircularProgress,
-  Divider,
   IconButton,
   Typography,
 } from '@mui/material'
 import useAxios from 'axios-hooks'
 import Breadcrumbs from 'components/Breadcrumbs'
+import { Divider } from 'components/Divider'
 import PageHeader from 'components/PageHeader'
 import RequestsTable from 'components/table'
 import { ServerConfigContainer } from 'containers/ConfigContainer'
