--- src/pages/RequestsIndex/RequestsIndexTable/index.ts
@@ -1,1 +1,0 @@
-export * from './RequestsIndexTable'
