+++ src/pages/RequestsIndex/data/dataHelpers.tsx
@@ -1,7 +1,7 @@
 import FileUploadIcon from '@mui/icons-material/FileUpload'
 import VisibilityOffIcon from '@mui/icons-material/VisibilityOff'
 import { Box, Tooltip, Typography } from '@mui/material'
-import { RequestsIndexTableData } from 'pages/RequestsIndex/RequestsIndexTable/data'
+import { RequestsIndexTableData } from 'pages/RequestsIndex/data'
 import { Link as RouterLink } from 'react-router-dom'
 import { Request } from 'types/backend-types'
