+++ src/pages/SystemIndex/SystemIndex.tsx
@@ -1,10 +1,22 @@
 import { Box } from '@mui/material'
-import { SystemIndexTable } from 'pages/SystemIndex/SystemIndexTable'
+import { Divider } from 'components/Divider'
+import PageHeader from 'components/PageHeader'
+import { Table } from 'components/Table'
+import {
+  useSystemIndexTableColumns,
+  useSystemIndexTableData,
+} from 'pages/SystemIndex'
 
 const SystemsIndex = () => {
   return (
-    <Box pt={1}>
-      <SystemIndexTable />
+    <Box>
+      <PageHeader title="Systems" description="" />
+      <Divider />
+      <Table
+        tableKey="Systems"
+        data={useSystemIndexTableData()}
+        columns={useSystemIndexTableColumns()}
+      />
     </Box>
   )
 }
