--- src/pages/SystemIndex/SystemIndexTable/index.ts
@@ -1,3 +1,0 @@
-export * from './data'
-export * from './ExploreButton'
-export * from './SystemIndexTable'
