--- src/pages/SystemIndex/SystemIndexTable/systemIndexHelpers.tsx
@@ -1,23 +1,0 @@
-import { System } from 'types/backend-types'
-
-import { ExploreButton } from '.'
-
-const formatSystems = (systems: System[]) => {
-  const formattedSystems: (string | JSX.Element | number)[][] = []
-
-  for (const index in systems) {
-    formattedSystems.push([
-      systems[index].namespace,
-      systems[index].name,
-      systems[index].version,
-      systems[index].description,
-      systems[index].commands.length,
-      systems[index].instances.length,
-      ExploreButton(systems[index]),
-    ])
-  }
-
-  return formattedSystems
-}
-
-export { formatSystems }
