+++ src/pages/SystemIndex/index.ts
@@ -1,1 +1,3 @@
+export * from './ExploreButton'
 export * from './SystemIndex'
+export * from './systemIndexHelpers'
